Added accuracy test function

+++ assignment2/src/assignment2/MatrixPac.java
@@ -3,6 +3,6 @@ package assignment2;
 //Hold 2d array of x and 1d matrix/array
 
 public class MatrixPac {
-	public int [][] x_values;
-	public int [] y_values;
+	public double [][] x_values;
+	public double [] y_values;
 }
